+++ samples/DLCBuilder.Domain/EditFunctions.fs
@@ -155,6 +155,9 @@ let editConfig edit config =
     | SetCharterName name ->
         { config with CharterName = name }
 
+    | SetReleaseFilenameAddition addition ->
+        { config with ReleaseFilenameAddition = Option.ofString addition }
+
     | SetAutoVolume autoVolume ->
         { config with AutoVolume = autoVolume }