+++ samples/DLCBuilder.Domain/StateTypes.fs
@@ -152,6 +152,7 @@ type VocalsEdit =
 
 type ConfigEdit =
     | SetCharterName of string
+    | SetReleaseFilenameAddition of string
     | SetProfilePath of string
     | SetTestFolderPath of string
     | SetDlcFolderPath of string